break wallet into several partial classes to prepare for further update.
